Merged in AL-63-vmcloak-and-building-vms-better (pull request #100)
AL-63 vmcloak and building vms better

Significant re-work for how VMs are expected to be created and imported into AL cuckoo service.

VM export/import. This is the main change for this PR.

Created ‘vmprep.py’ and ‘import-vm.py’. See README for usage details. We now rely on the third party ‘vmcloak’ tool to do create a headless ISO for installation, and this takes care of installing the cuckoo agent and ensuring that windows logs in without a user.
The goal was to make it much simpler to create and modify VMs, and make it simpler to import them into a cluster. import-vm.py will try and take care of copying all files to the right places for you.
Introduced the ‘analysis_vm’ submission parameter. This allows users to pick a specific VM to use for an analysis job
Deprecated the ‘vm_meta’ service configuration. The list of available VMs will be kept in the ‘analysis_vm' submission parameter. Instead we just directly read the _meta.json files
UPGRADING: If upgrading from an existing installation, cuckoo will continue to make use of the ‘vm_meta’ parameter as long as the ‘analysis_vm’ parameter doesn’t exist. If both exist, the service won’t start.
removed any references to ‘fakenet’ IP. As far as I could tell this was related to legacy functionality

no longer using RAM tmpfs to store a copy of the KVM snapshot - just copied locally into the docker container. OS filesystem cache should keep frequently used files in RAM. This should reduce RAM requirements if you want to have many different VMs available.